refactor(renderer): extract video source resolver
